"Update gradle wrapper and fix CDPATH issue
The gradle-wrapper version was updated from 8.2.1 to 8.3 to stay up-to-date with the latest stable release, which should potentially include improvements and bug fixes.

On the other hand, a change was made in the gradlew script to discard cd standard output. The modification ensures the script works as expected even when the CDPATH environment variable is set, solving a potential issue reported on GitHub (#25036)."